+++ src/app/(auth)/login/page.tsx
@@ -1,27 +1,15 @@
-"use client";
+import LoginContent from "@/components/authPage/LoginContent";
+import { Metadata } from "next";
 
-import LoginForm from "@/components/authPage/Form/LoginForm";
-import GithubButton from "@/components/common/button/GithubButton";
-import GoogleButton from "@/components/common/button/GoogleButton";
-import KakaoButton from "@/components/common/button/KakaoButton";
-import Image from "next/image";
-import LoginLogo from "@images/loginLogo.svg";
+export const metadata: Metadata = {
+  title: "로그인 페이지",
+  description: "이메일로 로그인하거나 소셜로그인을 할 수 있습니다."
+};
 
 const LoginPage = () => {
   return (
     <div className="-mt-3 flex h-screen flex-col items-center justify-center">
-      <div className="flex w-[20.9rem] flex-col items-center gap-[3.5rem] md:w-[27.5rem] lg:w-[32.5rem] lg:rounded-3xl lg:px-[2.5rem] lg:py-[3.88rem] lg:shadow-[0px_4px_20px_0px_rgba(154,130,102,0.1)]">
-        <Image src={LoginLogo} alt="집밥도감 로고" />
-        <LoginForm />
-        <div className="flex flex-col items-center">
-          <p className="text-body-14 text-[#777]">SNS로 로그인/회원가입</p>
-          <div className="mt-6 flex gap-[1.125rem]">
-            <GoogleButton />
-            <KakaoButton />
-            <GithubButton />
-          </div>
-        </div>
-      </div>
+      <LoginContent />
     </div>
   );
 };
